+++ legends/utils/__init__.py
@@ -12,6 +12,7 @@ Example:
 
 from legends.utils.functions import *
 from legends.utils.customabcs import *
+from legends.utils.htmltagstripper import *
 from legends.utils.relations import *
 from legends.utils.objrelations import *
 from legends.utils.scrollframe import *